moving: Modularized code; "State Machine" approach; Can cross screen borders
The code has been modified to migrate the previous approach to a state machine approach.
The migration is not yet complete.

The behavior of the player not being able to cross the screen borders has been
temporarily removed and will be reintroduced soon.

In addition, the code has been split into more files, which was necessary for the
introduction of the state machine.